% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_conv_one_to_three.R
\name{amino_acid_conversion_three_to_one}
\alias{amino_acid_conversion_three_to_one}
\title{\subsection{Change three to one letter code}{
}}
\usage{
amino_acid_conversion_three_to_one(protein_change_string)
}
\arguments{
\item{protein_change_string}{}
}
\description{
\subsection{Change three to one letter code}{
}
}
