% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_conv_one_to_three.R
\name{amino_code_conversion_one_to_three}
\alias{amino_code_conversion_one_to_three}
\title{Change amino acid one letter code to three letter code}
\usage{
amino_code_conversion_one_to_three(protein_change_string)
}
\arguments{
\item{protein_change_string}{}
}
\description{
Change amino acid one letter code to three letter code
}
\examples{
Changes p.Y772_A775dup to --> p.Tyr772_Ala775dup
}
